+++ renderTexture.go
@@ -168,7 +168,7 @@ func (this *RenderTexture) Draw(drawer Drawer, renderStates RenderStates) {
 func (this *RenderTexture) DrawPrimitives(vertices []Vertex, primType PrimitiveType, renderStates RenderStates) {
 	if len(vertices) > 0 {
 		rs := renderStates.toC()
-		C.sfRenderTexture_drawPrimitives(this.cptr, (*C.sfVertex)(unsafe.Pointer(&vertices[0])), C.uint(len(vertices)), C.sfPrimitiveType(primType), &rs)
+		C.sfRenderTexture_drawPrimitives(this.cptr, (*C.sfVertex)(unsafe.Pointer(&vertices[0])), C.size_t(len(vertices)), C.sfPrimitiveType(primType), &rs)
 	}
 }
